ajout du .env et ajout de celui-ci en gitignore, ajout de helmet et express-rate-limit, ajout de auth sur toutes les routes et condition supplémentaire dans la fonction deleteSauce

+++ backend/sauces/CRUD/sauce-delete.js
@@ -6,13 +6,17 @@ module.exports = () => {
   router.delete("/sauces/:id", (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
       .then((sauce) => {
-        const filename = sauce.imageUrl.split("assets/images/")[1];
-        fs.unlink(`assets/images/${filename}`, () => {
-          Sauce.deleteOne({ _id: req.params.id })
-            .then(() => res.status(200).json({ message: "Sauce supprimée" }))
-            .catch((error) => res.status(400).json({ error }));
-          console.log("Sauce supprimée");
-        });
+        if (sauce.userId != req.auth.userId) {
+          res.status(401).json({ message: "Not authorized" });
+        } else {
+          const filename = sauce.imageUrl.split("assets/images/")[1];
+          fs.unlink(`assets/images/${filename}`, () => {
+            Sauce.deleteOne({ _id: req.params.id })
+              .then(() => res.status(200).json({ message: "Sauce supprimée" }))
+              .catch((error) => res.status(400).json({ error }));
+            console.log("Sauce supprimée");
+          });
+        }
       })
       .catch((error) => res.status(500).json({ error }));
   });
